Refactor OtlpAspNetTestBase to move much of functionality to OtlpServerTestBase, so we can re-use it for testing the Owin.WebApi2 app
